Weight Initialization Updates
Fixed the bug "Boolean value of Tensor with more than one value is ambiguous" but may not be the most efficient way. I looped through each weight for each neuron in each layer as Noam suggested. I was more so just working out the syntax and seeing if this would work. There are still bugs in the forward function